fix: reuse burn tx logic
