+++ package.json
@@ -26,7 +26,7 @@
 		"@synthetixio/contracts-interface": "2.50.0-ovm-alpha",
 		"@synthetixio/optimism-networks": "2.48.4",
 		"@synthetixio/providers": "2.48.4",
-		"@synthetixio/queries": "2.50.0",
+		"@synthetixio/queries": "0.0.0-78f2647",
 		"@synthetixio/transaction-notifier": "2.48.4",
 		"@synthetixio/wei": "2.48.4",
 		"@tippyjs/react": "4.1.0",
